chore: add status call, move empty message to util
